+++ src/app/src/UI/base/style/AppBarStyle.ts
@@ -1,0 +1,25 @@
+import MuiAppBar, { AppBarProps as MuiAppBarProps } from "@mui/material/AppBar";
+import { styled } from "@mui/material/styles";
+
+
+interface AppBarProps extends MuiAppBarProps {
+  open?: boolean;
+  drawerWidth: number;
+}
+
+export const AppBarStyle = styled(MuiAppBar, {
+  shouldForwardProp: (prop) => prop !== "open" && prop !== "drawerWidth",
+})<AppBarProps>(({ theme, open, drawerWidth }) => ({
+  transition: theme.transitions.create(["margin", "width"], {
+    easing: theme.transitions.easing.sharp,
+    duration: theme.transitions.duration.leavingScreen,
+  }),
+  ...(open && {
+    width: `calc(100% - ${drawerWidth}px)`,
+    marginLeft: `${drawerWidth}px`,
+    transition: theme.transitions.create(["margin", "width"], {
+      easing: theme.transitions.easing.easeOut,
+      duration: theme.transitions.duration.enteringScreen,
+    }),
+  }),
+}));
