+++ src/app/src/UI/base/style/DrawerHeaderStyle.ts
@@ -1,0 +1,9 @@
+import { styled } from "@mui/material/styles";
+
+export const DrawerHeaderStyle = styled("div")(({ theme }) => ({
+  display: "flex",
+  alignItems: "center",
+  padding: theme.spacing(0, 1),
+  ...theme.mixins.toolbar,
+  justifyContent: "flex-end",
+}));
