+++ src/app/src/components/base/logout.ts
@@ -1,0 +1,11 @@
+import { auth } from "../../firebase";
+
+export const funcLogout = () => {
+  try {
+    auth.signOut();
+    return true
+  } catch (error) {
+    console.error(error);
+    return false
+  }
+}
